Detail Debt Completed

+++ money-management-booking/src/main/java/com/group6/moneymanagementbooking/controller/DebtorController.java
@@ -1,8 +1,7 @@
 package com.group6.moneymanagementbooking.controller;
 
-import java.time.LocalDate;
 import java.time.LocalDateTime;
-import java.util.Date;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -15,6 +14,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import com.group6.moneymanagementbooking.enity.Debtor;
@@ -25,33 +25,61 @@ import lombok.RequiredArgsConstructor;
 @Controller
 @RequiredArgsConstructor
 @CrossOrigin
+@RequestMapping("/Debtor")
 public class DebtorController {
     private final DebtorService debtorService;
 
-    @GetMapping("/ListDebtor")
+    @GetMapping("/ListAll")
     public String listDebtor(Model model) {
         model.addAttribute("list_debtor", debtorService.findAll());
         model.addAttribute("record", debtorService.findAll().size());
         return "view-debtor";
     }
 
-    @GetMapping("/AddDebtor")
+    @GetMapping("/ListDebtor")
+    public String AllDebtor(Model model) {
+        List<Debtor> listdeb = new ArrayList<>();
+        for (var deb : debtorService.findAll()) {
+            if (deb.getTotal() > 0) {
+                listdeb.add(deb);
+            }
+        }
+        model.addAttribute("list_debtor", listdeb);
+        model.addAttribute("record", listdeb.size());
+        return "view-debtor";
+    }
+
+    @GetMapping("/ListOwner")
+    public String AllOwner(Model model) {
+        List<Debtor> listdeb = new ArrayList<>();
+        for (var deb : debtorService.findAll()) {
+            if (deb.getTotal() < 0) {
+                listdeb.add(deb);
+            }
+        }
+        model.addAttribute("list_debtor", listdeb);
+        model.addAttribute("record", listdeb.size());
+        return "view-debtor";
+    }
+
+    @GetMapping("/Add")
     public String registerGet(Model model, HttpServletRequest request) {
         Debtor debtor = new Debtor();
         model.addAttribute("debtor", debtor);
-         model.addAttribute("title", "Add");
+        model.addAttribute("title", "Add");
         return "add-debtor";
     }
 
-    @PostMapping("/Add-Debtor")
+    @PostMapping("/Add")
     public String registerPost(Model model, @ModelAttribute("debtor") Debtor debtor) throws Exception {
+        debtor.setTotal(0.0);
         debtor.setDate_create(LocalDateTime.now());
         debtor.setDate_update(LocalDateTime.now());
         debtorService.Save(debtor);
-        return "redirect:/ListDebtor";
+        return "redirect:/Debtor/ListAll";
     }
 
-    @PostMapping("/Search-Debtor")
+    @PostMapping("/Search")
     public String searchDebtor(Model model, @RequestParam("nameDebtor") String name) throws Exception {
         model.addAttribute("list_debtor", debtorService.SearchByName(name));
         model.addAttribute("record", debtorService.SearchByName(name).size());
@@ -59,19 +87,25 @@ public class DebtorController {
         return "view-debtor";
     }
 
-    @GetMapping("/edit-Debtor/{id}")
+    @GetMapping("/edit/{id}")
     public String registerGet(Model model, @PathVariable("id") int id) {
-        Optional<Debtor>  debtor = debtorService.getDebtor(id);
+        Optional<Debtor> debtor = debtorService.getDebtor(id);
         model.addAttribute("debtor", debtor.get());
-          model.addAttribute("title", "Update");
+        model.addAttribute("title", "Update");
         return "add-debtor";
     }
 
-     @GetMapping("/delete-Debtor/{id}")
+    @GetMapping("/delete/{id}")
     public String deleteDebtor(Model model, @PathVariable("id") int id) {
-         debtorService.deleteDebtorById(id);
-        return "redirect:/ListDebtor";
+        Debtor debtor = (debtorService.getDebtor(id)).get();
+        if (debtor.getTotal() != 0) {
+            model.addAttribute("list_debtor", debtorService.findAll());
+            model.addAttribute("record", debtorService.findAll().size());
+            model.addAttribute("mess", "Cannot be delete because the debt has not been completed!");
+            return "view-debtor";
+        }
+        debtorService.deleteDebtorById(id);
+        return "redirect:/Debtor/ListAll";
     }
 
-   
 }
